+++ SandboxiePlus/SandMan/Windows/OptionsAdvanced.cpp
@@ -106,6 +106,8 @@ void COptionsWindow::CreateAdvanced()
 	InitLangID();
 
 	connect(ui.chkHideFirmware, SIGNAL(clicked(bool)), this, SLOT(OnAdvancedChanged()));
+	connect(ui.chkHideUID, SIGNAL(clicked(bool)), this, SLOT(OnAdvancedChanged()));
+	connect(ui.chkHideSerial, SIGNAL(clicked(bool)), this, SLOT(OnAdvancedChanged()));
 	connect(ui.cmbLangID, SIGNAL(currentIndexChanged(int)), this, SLOT(OnAdvancedChanged()));
 	connect(ui.btnDumpFW, SIGNAL(clicked(bool)), this, SLOT(OnDumpFW()));
 
@@ -289,6 +291,8 @@ void COptionsWindow::LoadAdvanced()
 	//
 
 	ui.chkHideFirmware->setChecked(m_pBox->GetBool("HideFirmwareInfo", false));
+	ui.chkHideUID->setChecked(m_pBox->GetBool("RandomRegUID",false));
+	ui.chkHideSerial->setChecked(m_pBox->GetBool("HideDiskSerialNumber", false));
 
 	ui.cmbLangID->setCurrentIndex(ui.cmbLangID->findData(m_pBox->GetNum("CustomLCID", 0)));
 
@@ -575,6 +579,8 @@ void COptionsWindow::SaveAdvanced()
 	//
 
 	WriteAdvancedCheck(ui.chkHideFirmware, "HideFirmwareInfo", "y", "");
+	WriteAdvancedCheck(ui.chkHideUID, "RandomRegUID", "y", "");
+	WriteAdvancedCheck(ui.chkHideSerial, "HideDiskSerialNumber", "y", "");
 
 	int CustomLCID = ui.cmbLangID->currentData().toInt();
 	if (CustomLCID) m_pBox->SetNum("CustomLCID", CustomLCID);